+++ dev/e2e_app/macos/Runner/AppDelegate.swift
@@ -3,6 +3,13 @@ import FlutterMacOS
 
 @main
 class AppDelegate: FlutterAppDelegate {
+  private(set) static var patrolMenuActionCount = 0
+
+  @IBAction
+  func patrolMenuAction(_ sender: Any?) {
+    Self.patrolMenuActionCount += 1
+  }
+
   override func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {
     return true
   }